# openapi.yaml
openapi: 3.1.0
info:
  title: Herald API
  description: Persistent realtime event streams with built-in authorization.
  license:
    name: Proprietary
  version: 2.0.0
servers:
- url: /
  description: Local server
paths:
...
  /blocks:
    post:
      tags:
      - blocks
      operationId: block_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockRequest'
        required: true
      responses:
        '201':
          description: User blocked
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
    delete:
      tags:
      - blocks
      operationId: unblock_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockRequest'
        required: true
      responses:
        '204':
          description: User unblocked
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Block not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
  /blocks/{user_id}:
    get:
      tags:
      - blocks
      operationId: list_blocked
      parameters:
      - name: user_id
        in: path
        description: User ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: List of blocked users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockListResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
...
  /streams/{id}/cursors:
    get:
      tags:
      - cursors
      operationId: list_cursors
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Read cursors for stream
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorListResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
...
  /streams/{id}/events/{event_id}:
    delete:
      tags:
      - events
      operationId: delete_event
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      - name: event_id
        in: path
        description: Event ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Event deleted
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
    patch:
      tags:
      - events
      operationId: edit_event
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      - name: event_id
        in: path
        description: Event ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditEventRequest'
        required: true
      responses:
        '200':
          description: Event edited
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
  /streams/{id}/events/{event_id}/reactions:
    get:
      tags:
      - events
      operationId: get_reactions
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      - name: event_id
        in: path
        description: Event ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Reactions for event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionListResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
...
components:
  schemas:
...
    BlockListResponse:
      type: object
      required:
      - blocked
      properties:
        blocked:
          type: array
          items:
            type: string
    BlockRequest:
      type: object
      required:
      - user_id
      - blocked_id
      properties:
        blocked_id:
          type: string
        user_id:
          type: string
...
    CursorListResponse:
      type: object
      required:
      - cursors
      properties:
        cursors:
          type: array
          items:
            $ref: '#/components/schemas/CursorView'
    CursorView:
      type: object
      required:
      - user_id
      - seq
      properties:
        seq:
          type: integer
          format: int64
          minimum: 0
        user_id:
          type: string
    EditEventRequest:
      type: object
      required:
      - body
      properties:
        body:
          type: string
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    ReactionListResponse:
      type: object
      required:
      - reactions
      properties:
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/ReactionSummary'
    ReactionSummary:
      type: object
      required:
      - emoji
      - count
      - users
      properties:
        count:
          type: integer
          minimum: 0
        emoji:
          type: string
        users:
          type: array
          items:
            type: string
...
  securitySchemes:
    admin_auth:
      type: http
      scheme: bearer
    basic_auth:
      type: http
      scheme: basic
    bearer_auth:
      type: http
      scheme: bearer
security:
- basic_auth: []
- bearer_auth: []
- admin_auth: []
...